add to full section 25

+++ app/Http/Controllers/PostsController.php
@@ -6,9 +6,17 @@ use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class PostsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')
+            ->only(['create', 'store', 'edit', 'update', 'destroy']);
+    }
+
+
     /**
      * Display a listing of the resource.
      *
@@ -53,17 +61,24 @@ class PostsController extends Controller
      */
     public function store(StorePost $request)
     {
-        $validated = $request->validated();
+        // $validated = $request->validated();
 
-        $post = new BlogPost();
+        // $post = new BlogPost();
 
-        $post->title = $validated['title'];
-        $post->content = $validated['content'];
-        $post->save();
+        // $post->title = $validated['title'];
+        // $post->content = $validated['content'];
+        // $post->save();
 
-        $request->session()->flash('status', 'The blog post was created!');
+        // $request->session()->flash('status', 'The blog post was created!');
 
-        return redirect()->route('posts.show', ['post' => $post->id]);
+        // return redirect()->route('posts.show', ['post' => $post->id]);
+
+
+        $validatedData = $request->validated();
+        $blogPost = BlogPost::create($validatedData);
+        $request->session()->flash('status', 'Blog post was created!');
+
+        return redirect()->route('posts.show', ['post' => $blogPost->id]);
     }
 
     /**
@@ -85,7 +100,16 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
-        return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
+        // return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
+
+        $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog post");
+        // }
+        $this->authorize('posts.update', $post);
+
+        return view('posts.edit', ['post' => $post]);
     }
 
     /**
@@ -98,6 +122,11 @@ class PostsController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog post");
+        // }
+        $this->authorize('posts.update', $post);
 
         $validated = $request->validated();
         $post->fill($validated);
@@ -115,12 +144,18 @@ class PostsController extends Controller
      * @param  \App\Models\BlogPost  $blogPost
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't delete this blog post");
+        // }
+        $this->authorize('posts.delete', $post);
+
         $post->delete();
 
-        session()->flash('status', 'Blog post was deleted!');
+        $request->session()->flash('status', 'Blog post has been deleted!');
 
         return redirect()->route('posts.index');
     }
